requirements.txt
pandas 
notebook
numpy
scikit-learn
sklearn_features
imblearn
matplotlib
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.4
joblib
types-PyYAML
Flask
Flask-Cors
-e .